Pi Code/requirements.txt
Adafruit-Blinka==8.54.0
adafruit-circuitpython-bme680==3.7.12
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-requests==4.1.9
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
bme680==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pdoc==15.0.1
pvporcupine==3.0.5
pvrecorder==1.2.5
PyAudio==0.2.14
pyftdi==0.56.0
Pygments==2.19.1
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
smbus2==0.5.0
sysv-ipc==1.1.0
toml==0.10.2
typing_extensions==4.12.2
